push stuff so i can actually create a branch instead

+++ game/camera.py
@@ -1,5 +1,5 @@
 class Camera:
-    pos: tuple[int, int]
+    pos: tuple[float, float]
 
-    def __init__(self, pos: tuple[int, int]):
+    def __init__(self, pos: tuple[float, float]):
         self.pos = pos
